Added bitstore - not quite working yet.

+++ drivers/bitstore.h
@@ -2,7 +2,7 @@
 #define _BITSTORE_H
 
 int bitstore_GetBlock(uint8_t *blk);
-int bitstore_Store(void *user_data, int (*get_block)(void *user_data, uint8_t *block));
+int bitstore_Store(void *user_data, uint8_t (*get_block)(void *user_data, uint8_t *block));
 int bitstore_Size();
 int bitstore_InitRetrieve();
 void bitstore_Free();
